Implement comprehensive logging framework for RedisCC
- Added log management functionalities in `log.cc`, including initialization, shutdown, context management, and statistics retrieval.
- Developed `logger_config.cc` to handle logger configuration, including loading from environment variables and JSON files, with default settings and validation.
- Created `logger_manager.cc` to manage logger instances, including asynchronous logging setup, sink creation, and context handling.
- Introduced utility functions in `logger_utils.cc` for scoped logging, performance tracking, request tracking, and log statistics collection.
- Added a suppression file for Valgrind to handle known false positives and third-party library issues.

+++ tests/format_test.cc
@@ -4,7 +4,7 @@ using namespace std;
 
 // 函数：打印一个整数数组
 void printArray( const vector< int >& arr ) {
-  for ( int i = 0; i < arr.size(); i++ ) {
+  for ( size_t i = 0; i < arr.size(); i++ ) {
     cout << arr[ i ] << " ";
   }
   cout << endl;
@@ -13,7 +13,7 @@ void printArray( const vector< int >& arr ) {
 // 函数：计算数组中所有元素的和
 int sumArray( const vector< int >& arr ) {
   int sum = 0;
-  for ( int i = 0; i < arr.size(); i++ ) {
+  for ( size_t i = 0; i < arr.size(); i++ ) {
     sum += arr[ i ];
   }
   return sum;
